+++ main.h
@@ -88,7 +88,7 @@ typedef struct Entity {
 	ivec2s pos;
 	vec2s pos_remainder;
 	vec2s vel;
-	float dir;
+	int32_t dir;
 	Anim anim;
 	int32_t touching_floor;
 	EntityType type;
@@ -183,10 +183,10 @@ void SetSpriteFromPath(Entity* entity, const char* path);
 bool SetSprite(Entity* entity, Sprite sprite);
 SpriteDesc* GetSpriteDesc(Context* ctx, Sprite sprite);
 void LoadSprite(SDL_Renderer* renderer, SDL_IOStream* fs, SpriteDesc* sd);
-void DrawSprite(Context* ctx, Sprite sprite, size_t frame_idx, ivec2s pos, float dir);
+void DrawSprite(Context* ctx, Sprite sprite, size_t frame_idx, ivec2s pos, int32_t dir);
 void DrawSpriteTile(Context* ctx, Sprite sprite, ivec2s tile, ivec2s ipos);
 void DrawEntity(Context* ctx, Entity* entity);
-void DrawAnim(Context* ctx, Anim* anim, ivec2s pos, float dir);
+void DrawAnim(Context* ctx, Anim* anim, ivec2s pos, int32_t dir);
 void UpdateAnim(Context* ctx, Anim* anim, bool loop);
 bool SpritesEqual(Sprite a, Sprite b);
